line plot
plus also did some research on things to do and visualization tools.
Seaborn is best for us I think and will try and do more with that.
I think we probably have time and scope to do some level of more complex text analysis, so will have a go at that.